Home components refactor
