+++ src/HomePage.jsx
@@ -1,6 +1,6 @@
-import Inquiry from "./components/Inquiry";
-import Banner from "./components/Banner";
-import FeaturedProducts from "./components/FeaturedProducts";
+import Inquiry from "./components/Home/Inquiry";
+import Banner from "./components/Home/Banner";
+import FeaturedProducts from "./components/Home/FeaturedProducts";
 import HomeLogin from "./components/HomeLogin";
 
 function HomePage({ menuData }) {
